test: 보너스 번호 중복 검증 테스트 WinningNumberTest로 변경
