+++ src/main/java/lotto/model/WinningNumbers.java
@@ -11,6 +11,9 @@ public class WinningNumbers {
         this.winningNumbers = winningNumbers;
     }
 
+    public void checkDuplicate(int bonusNumber) {
+    }
+
     private void validate(List<Integer> winningNumbers) {
         if (winningNumbers.size() != 6) {
             throw new IllegalArgumentException("[ERROR] 당첨 번호는 6개여야 합니다.");
